+++ packages/virtual/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @af-utils/virtual-core
 
+## 1.0.2
+
+### Patch Changes
+
+- upgraded dependencies, added float dimensions support
+
 ## 1.0.1
 
 ### Patch Changes